==> notebooks/01_data_preprocessing.ipynb <==
import pandas as pd
import config

pd.set_option('display.max_rows',100)

### Load and process trips data

trips = pd.read_csv(config.TRIPS_1216_RAW, low_memory=False)
trips.shape

trips.head(5).T

trips.Mode1.value_counts()

trips.LINKMODE.value_counts()

trips.columns

### Select related columns

columns = ['TRIPID', 'PERSID', 'TRIPNO', 'TRAVDOW', 'LINKMODE', 'TRIPPURP', 'DESTPURP1',
           'STARTHOUR', 'STARTIME', 'ARRHOUR', 'ARRTIME', 'TRIPTIME', 'TRAVTIME', 'WAITIME', 
           'ORIGSA1', 'ORIGSA3', 'DESTSA1', 'DESTSA3', 'CUMDIST']

trips = trips[columns]

trips.LINKMODE.value_counts()

### Select trips with Start_Hour between 0h-24h

trips = trips[trips['STARTHOUR'] < 24]
trips.shape

trips['STARTHOUR'].value_counts()

### Process main travel mode from LinkMode

trips.LINKMODE.value_counts()

# Map to 4 main travel modes: Walk, Car, Train, CycleTRIPPURP
mode_mapping = {'Walking':'Walk', 'Vehicle Driver': 'Car', 'Vehicle Passenger': 'Car', 'Taxi': 'Car',
                'Bicycle':'Cycle', 'Train':'Train', 'Tram':'Train', 'Public Bus': 'Car',}

trips['Mode'] = trips.LINKMODE.map(mode_mapping)

trips['Mode'].value_counts()

### Process trip purpose

trips.DESTPURP1.value_counts()

trips.TRIPPURP.value_counts()

# Map trip purpose
mapping = {"Work Related":"Work", "Buy Something":"Shopping", "Social":"Social", "Pick-up or Drop-off Someone":"Pick-up_Drop-off",
           "Recreational":"Recreational","Personal Business":"Personal Business","Education":"Education","Accompany Someone":"Pick-up_Drop-off",
           "Pick-up or Deliver Something":"Pick-up_Drop-off", "Other Purpose":"Other", "Unknown purpose (at start of day)":"Other",
           "Change Mode":"Other", "At or Go Home":"Other", "Not Stated":"Other"}
trips.TRIPPURP = trips.TRIPPURP.map(mapping)

trips.TRIPPURP.value_counts()

### Select people with at least one trip with active mode (Walking, Bicycle)

# How many people have at leat one active mode trip
trips[trips.LINKMODE.isin(['Walking','Bicycle'])].PERSID.nunique()

active_persons = trips[trips.LINKMODE.isin(['Walking','Bicycle'])].PERSID.unique()
type(active_persons), len(active_persons)

# Get trips of these active persons
trips_ap = trips[trips.PERSID.isin(active_persons)]
trips_ap.shape

trips_ap.isna().sum()

trips_ap = trips_ap.dropna()

trips_ap.shape

### Save to csv file for further analysis

trips_ap['Mode'].value_counts()

trips_ap.to_csv(config.TRIPS_1216_PRC, index=False)

trips_ap.columns

trips_ap.head()
