==> notebooks/03_los_goolge_api_utils.ipynb <==
## OD Level of sevice from Google API
This notebook download trip travel time for different mode (car, public transport, cycling, walking) using Google API

https://cloud.google.com/maps-platform/pricing

Get USD 200.00 in free usage for Maps, Routes, and Places every month => CAN GET AROUND 30k request/month FOR FREE


* TIER 1: AUD 6.44525/1K requests FOR 0 - 100K requests/month
* TIER 2: AUD 5.1562/1K requests  FOR 100K+ requests/month

#export
import os
import googlemaps
from datetime import datetime

#export
# Setup google elevation api key
GCP_MAP_KEY = os.environ.get('GCP_ELEVATION_API')
gmaps = googlemaps.Client(key=GCP_MAP_KEY)

### Transit mode example

# Request directions via public transit
now = datetime.now()
directions_result = gmaps.directions("Monash University, Clayton, Vic",
                                     "Melbourne Museum, Vic",
                                     mode="transit",
                                     departure_time=now)

# get_direction_info(directions_result) => RETURN total distance and total duration INCLUDE walking travel.
#    'distance': {'text': '22.1 km', 'value': 22148},
#    'duration': {'text': '1 hour 0 mins', 'value': 3602},

directions_result[0]

### Gmap direction APIs: Driving

now = datetime.now()
now

orig = [-37.89577598106494, 145.09038583978858]
dest = [-37.91291561061539, 145.13363175213212]
modes = ["driving", "walking", "bicycling", "transit"]
mode = "driving"

distance = gmaps.directions(orig, dest,mode=mode, departure_time=now)

direction = distance
type(direction)

direction

direction[0]['legs'][0]['distance']['value'], direction[0]['legs'][0]['duration']

### Get LoS from Google Direction APIs

#export

def get_direction_info(direction):
    try:
        travel_dur = direction[0]['legs'][0]['duration']['value']
        travel_len = direction[0]['legs'][0]['distance']['value']
        return [travel_dur, travel_len]
    except:
        return [0, 0]

def collect_google_los(gmap, orig, dest, mode, departure_time=None):
    """
    This function collect level of service (travel time) from HERE APIs
    
    Parameters
        gmap: googlemaps Client object
        orig: origin location as a [lattitue, longtitude] list
        dest: destination location as a [lattitue, longtitude] list
        mode: travel mode like 'transit', 'driving', 'walking', 'bicycling'
        departure_time: the trip's departure time in datetime format
    Return
        duration (senconds), len (meter)
    """    
    direction = gmap.directions(orig, dest, mode=mode, departure_time=departure_time)
    res = get_direction_info(direction)
    return res

# Testing
orig = [-37.89577598106494, 145.09038583978858]
dest = [-37.91291561061539, 145.13363175213212]
mode = "transit"
departure_time = datetime.strptime("27/08/2021 23:30",'%d/%m/%Y %H:%M')
travel_info = collect_google_los(gmaps, orig, dest, mode, departure_time)

travel_info

### Convert notebook to python file

!python nb2scripts.py 03_los_goolge_api_utils.ipynb
