==> notebooks/04_los_here_api_utils.ipynb <==
### Using HERE APIs to get travel time and distance
HERE FREE ACCOUNT

https://developer.here.com/pricing

Free Transactions per month 250K

Pay per additional Transactions 1000 / $1

#export
import json
import os
from datetime import datetime
from here_location_services import LS
from here_location_services.config.routing_config import ROUTING_RETURN

#export
# Get HERE api object
LS_API_KEY = os.environ.get("HERE_API_KEY")  # Get API KEY from environment.
ls = LS(api_key=LS_API_KEY)

### HERE geocoding

address = "Monash University, Clayton, Vic"
geo = ls.geocode(query=address)
geo

print(geo)

### HERE routing APIs: Car Driving

Respone data type: https://developer.here.com/documentation/routing/dev_guide/topics/resource-type-route-summary.html

orig = [-37.89577598106494, 145.09038583978858]
dest = [-37.91291561061539, 145.13363175213212]

result = ls.car_route(
    origin=orig,
    destination=dest,
    return_results=[
#         ROUTING_RETURN.polyline,
#         ROUTING_RETURN.elevation,
#         ROUTING_RETURN.instructions,
#         ROUTING_RETURN.actions,
        ROUTING_RETURN.summary
    ],
)

print(result)

# Using trip start_time
start_time = datetime.strptime("27/05/2012 23:30",'%d/%m/%Y %H:%M')
result = ls.car_route(
    origin=orig,
    destination=dest,
    departure_time=start_time,
    return_results=[ ROUTING_RETURN.summary ],
)

RoutingResponse??

result??

print(result)

#export

def get_info(los_res):
    """
    This function parse result from HERE route api, and return travel_time and trip length
    Parameters
        los_res: result from HERE rout api
    Return
        (trip_duration in seconds, trip_length in meters)
    """
    try:
        res = los_res.routes[0]['sections'][0]['summary']
        return res['duration'], res['length']
    except:
        return 0, 0

def collect_here_los(ls, orig, dest, departure_time=None):
    """
    This function collect level of service (travel time) from HERE APIs
    
    Parameters
        ls: HERE location services (LS) object
        orig: origin location as a [lattitue, longtitude] list
        dest: destination location as a [lattitue, longtitude] list
        departure_time: the trip's departure time in datetime format
    Return
        [car_duration, car_length, cycle_duration, cycle_length, walk_duration, walk_length]
    """
    try:
        # Get car duration and length
        car_res = ls.car_route(origin=orig, destination=dest, departure_time=departure_time, 
                            return_results=[ROUTING_RETURN.summary])
        car_dur, car_len = get_info(car_res)
        # Get cycle duration and length
        cycle_res = ls.bicycle_route(origin=orig, destination=dest, departure_time=departure_time, 
                            return_results=[ROUTING_RETURN.summary])
        cycle_dur, cycle_len = get_info(cycle_res)
        # Get walk duration and length
        walk_res = ls.pedestrian_route(origin=orig, destination=dest, departure_time=departure_time, 
                            return_results=[ROUTING_RETURN.summary])
        walk_dur, walk_len = get_info(walk_res)
        return [car_dur, car_len, cycle_dur, cycle_len, walk_dur, walk_len]    
    except:
        return [0, 0, 0, 0, 0, 0]  

#test
get_info(result)

#test
collect_here_los(ls, orig, dest, start_time)

### Convert notebook to python file

!python nb2scripts.py 04_los_here_api_utils.ipynb
