==> notebooks/05_preparing_trip_od_locations.ipynb <==
#### This notebook add centroid for origin and destination zone of trips
* The centroids of zones are generated from SA1 shapefile of ABS census
* 


import pandas as pd
import geopandas as gpd
import matplotlib.pyplot as plt
import config

trips = pd.read_csv(config.TRIPS_1216_PRC, low_memory=False)
trips.shape

trips.head()

trips.columns

### Get SA1 shapefile

shape_file = "../data/raw/1270055001_sa1_2016_aust_shape.zip"
gdf = gpd.read_file(shape_file)

gdf.shape

gdf[gdf['SA1_MAIN16'].isin([20604112201])]

gdf.head()

gdf.STE_NAME16.value_counts()

# Only using SA1 from Victoria
# gdf_vic = gdf[gdf.STE_NAME16 == "Victoria"]
gdf_vic = gdf
gdf_vic.shape

gdf_vic.columns

# Remove SA1 do not have geometry
gdf_vic.dropna(inplace=True)

gdf_vic.isna().sum()

columns = ['SA1_MAIN16', 'SA2_MAIN16', 'SA2_NAME16', 
           'SA3_CODE16', 'SA3_NAME16', 'SA4_CODE16', 'SA4_NAME16', 
           'STE_NAME16', 'AREASQKM16', 'geometry']
gdf_vic = gdf_vic[columns]
gdf_vic.shape

gdf_vic["centroid"] = gdf_vic.geometry.apply(lambda x: x.centroid)

gdf_vic.head()

### Merge OD SA1 centroids to trips dataset

trips.head()

trips = trips.merge(gdf_vic[['SA1_MAIN16','centroid']], how="left", left_on="ORIGSA1", right_on="SA1_MAIN16")
trips = trips.rename(columns={"centroid":"ORIG_CENTROID"})
trips = trips.drop(columns=["SA1_MAIN16"], axis=1)

trips = trips.merge(gdf_vic[['SA1_MAIN16','centroid']], how="left", left_on="DESTSA1", right_on="SA1_MAIN16")
trips = trips.rename(columns={"centroid":"DEST_CENTROID"})
trips = trips.drop(columns=["SA1_MAIN16"], axis=1)

trips.head()

trips.isna().sum()

trips[trips.isna()].shape

trips[trips.isna()].head()

trips = trips.dropna()

#### Get trip ORIG_XY and DEST_XY

trips.loc[2,].ORIG_CENTROID.xy

def get_x(trip, column):
    return trip[column].x

def get_y(trip, column):
    return trip[column].y

get_xy(trips.loc[0,], 'ORIG_CENTROID')

trips['ORIG_X'] = trips.apply(lambda row: get_x(row, 'ORIG_CENTROID'), axis=1)
trips['ORIG_Y'] = trips.apply(lambda row: get_y(row, 'ORIG_CENTROID'), axis=1)
trips['DEST_X'] = trips.apply(lambda row: get_x(row, 'DEST_CENTROID'), axis=1)
trips['DEST_Y'] = trips.apply(lambda row: get_y(row, 'DEST_CENTROID'), axis=1)

trips.TRAVDOW.value_counts()

### Saves trips OD locations into CSV

trips.to_csv("../data/processed/trip_od_locations.csv", index=False)

trips.shape

### Get person and household main info

##### Get household info

hh = pd.read_csv("../data/raw/VISTA_2012_16_v1_SA1_CSV/H_VISTA12_16_SA1_V1.csv")
hh.shape

cols = ['HHID', 'HHSIZE', 'VISITORS', 'DWELLTYPE',
       'OWNDWELL', 'YEARSLIVED', 'MONTHSLIVED', 'ADULTBIKES', 'KIDSBIKES',
       'TOTALBIKES', 'PHONE', 'CARS', 'FOURWDS', 'UTES', 'VANS', 'TRUCKS',
       'MBIKES', 'OTHERVEHS', 'TOTALVEHS', 'TRAVDATE', 'TRAVMONTH', 'TRAVYEAR',
       'HOMETAZ', 'YOUNGEST', 'AVEAGE', 'OLDEST', 'HHINC']

hh = hh[cols]
hh.shape

##### Get person info

ps = pd.read_csv("../data/raw/VISTA_2012_16_v1_SA1_CSV/P_VISTA12_16_SA1_V1.csv")
ps.shape

cols =['PERSID', 'HHID', 'PERSNO', 'WEEK', 'MONTHOFBIRTH',
       'YEAROFBIRTH', 'AGE', 'AGEGROUP', 'SEX', 'RELATIONSHIP', 'CARLICENCE',
       'MBIKELICENCE', 'OTHERLICENCE', 'NOLICENCE', 'FULLTIMEWORK',
       'PARTTIMEWORK', 'CASUALWORK', 'ANYWORK', 'STUDYING', 'ACTIVITIES',
       'MAINACT', 'WORKTYPE', 'EMPTYPE', 'ANZSCO1', 'ANZSIC1',
       'STARTPLACE', 'ANYSTOPS', 'NUMSTOPS', 'REASONCODE', 'MRTDOW',
       'MRTINT', 'PERSINC', 'WHOFILLED', 'PROXY', 'FILLDOW', 'FILLLAG',
       'ADDITIONALTRAVEL', 'NEWSTOP', 'DISABLEDTRANS', 'TAXITRANS',
       'WHEELCHAIRTRANS', 'ACCESSTRANS', 'NONETRANS', 'NOTTOSAYTRANS',
       'CYCLEDWORK', 'CYCLEDSHOPPING', 'CYCLEDEXERCISE', 'CYCLEDOTHER',
       'NOCYCLED', 'HOMESLA', 'HOMELGA', 'HomeSA1', 'HomeSA2', 'HomeSA3',
       'HomeSA4', 'HomeRegion_ASGS', 'HomeSubRegion_ASGS', 'AgeGroup_RW', 'STOPS',
       'ReportingPeriod']

ps = ps[cols]
ps.shape

##### Merge household info into person table

ps = ps.merge(hh, how='left', on='HHID')
ps.shape

ps.to_csv("../data/processed/persons.csv", index=False)

### Merge person info into Trip table

trips = trips.merge(ps, how="left", on='PERSID')
trips.shape

trips.to_csv("../data/processed/trip_od_full.csv", index=False)
