==> notebooks/07_gmap_los_collection.ipynb <==
### README: This notebook uses Google Map Direction APIs to collect the level of service for vista trips with following travel modes:
* transit

Note that we can also collect for other modes, but due to quote limitation, we only collect for `transit`. For other modes, we will use HERE APIs

%load_ext autoreload
%autoreload 2

import json, os, time
import numpy as np
import pandas as pd
import googlemaps
from datetime import datetime, timedelta
from los_goolge_api_utils import *

# Setup google api key
GCP_MAP_KEY = os.environ.get('GCP_ELEVATION_API')
gmap = googlemaps.Client(key=GCP_MAP_KEY)

### Load trips data

trips = pd.read_csv("../data/processed/trip_od_full.csv", low_memory=False)
trips.shape

# Use relavent columns
cols = ['TRIPID', 'Mode', 'LINKMODE', 'TRIPPURP', 'DESTPURP1', 'TRIPTIME', 'TRAVTIME',
        'TRAVDOW', 'TRAVDATE', 'TRAVMONTH', 'TRAVYEAR', 'STARTHOUR', 'STARTIME', 
        'HomeSA1', 'CUMDIST','ORIGSA1', 'DESTSA1', 'ORIG_X', 'ORIG_Y', 'DEST_X', 'DEST_Y']
trips = trips[cols]
trips.shape

### Get trips departure time

# Get detailed start time of trips
trips['STARTMIN'] = trips['STARTIME'] % 60
trips['START_DETAIL_TIME'] = (trips['TRAVDATE'].astype(str) + " " + trips['STARTHOUR'].astype(str) 
                              + ":" + trips['STARTMIN'].astype(str))

trips.head().T

### Collect Level of Service from Google APIs

def get_datetime(str_time):
    return datetime.strptime(str_time,'%d/%m/%Y %H:%M')

# Test
print(get_datetime("27/05/2012 23:30"))

def get_transit_future_departure_time(departure_time):
    """
    This function will get the transit future time for the same date of week of trip departure time,
    but in 2 weeks in futures.
    """
    tripday = get_datetime(departure_time)
    today = datetime.now()
    monday = today - timedelta(days=today.weekday())
    # Get same week day of the trip date
    sameday_thisweek = monday + timedelta(tripday.weekday())
    sameday_thisweek = sameday_thisweek.replace(hour=tripday.hour, minute=tripday.minute, second=tripday.second)
    # Get same day but in 2 weeks in future
    sameday_next2week = sameday_thisweek + timedelta(days=2*7) # Next 2 weeks 
    return sameday_next2week

# Test
dep_time = "06/08/2015 23:30"
get_transit_future_departure_time(dep_time)

# Test with trips dataframe
trips.head(20).START_DETAIL_TIME.apply(lambda x: get_transit_future_departure_time(x))

def get_trip_los(trip, mode='transit'):
    time.sleep(0.3)
    start_time = get_transit_future_departure_time(trip['START_DETAIL_TIME'])    
    orig = [trip.ORIG_Y, trip.ORIG_X]
    dest = [trip.DEST_Y, trip.DEST_X]
    (trip['TRANSIT_TIME'], trip['TRANSIT_LEN']) = collect_google_los(gmap, orig, dest, mode, start_time)
    return trip

%%time
# Test
trips.head(20).apply(lambda x: get_trip_los(x), axis=1)

trips.Mode.value_counts()

# Divide data into walk_trips and not_walk_trips for smaller APIs number of queries
car_trips = trips[trips.Mode.isin(['Car'])]
walk_trips = trips[trips.Mode.isin(['Walk'])]
train_trips = trips[trips.Mode.isin(['Train'])]
cycle_trips = trips[trips.Mode.isin(['Cycle'])]
assert len(trips) == len(car_trips) + len(walk_trips) + len(train_trips) + len(cycle_trips)
walk_trips.shape, car_trips.shape, train_trips.shape, cycle_trips.shape, trips.shape

# Use batch processing as limitation of Google Map APIs service
# ApiError: 429, Too Many Requests, {"error":"Too Many Requests","error_description":"Rate limit for this service has been reached"}

def process_trips_los(trips_df, batch_size, save_path, df_name):
    """
    Process trips and get los from Google Map APIs
    Parameters
        trips_df: trips dataframe
        batch_size: process in batch due to quota limitation of Google Map APIs
        save_path: where to save collected batch files
    """
    batch_idxs = np.arange(len(trips_df)) // batch_size
    for gr, df in trips_df.groupby(batch_idxs):        
        trip_los = df.apply(lambda x: get_trip_los(x), axis=1)
        trip_los.to_csv(f"{save_path}/{df_name}_{gr}.csv")
        print(f"Finished batch {gr+1}", df.shape, trip_los.shape)
        time.sleep(5)
        
        
def join_trips_los(trips_len, batch_size, save_path, df_name):
    batch_idxs = np.arange(trips_len // batch_size + 1)
    trips = []
    for idx in batch_idxs:
        trip_los = pd.read_csv(f"{save_path}/{df_name}_{idx}.csv")
        trips.append(trip_los)
    return pd.concat(trips)

%%time
# RUN ONCE - Get LOS for train_trips
save_path = "../data/los_gmap"
batch_size = 200
process_trips_los(train_trips, batch_size, save_path, "train_trips_los_gmap")

%%time
# RUN ONCE - Get LOS for car_trips
process_trips_los(car_trips, batch_size, save_path, "car_trips_los_gmap")

%%time
# RUN ONCE - Get LOS for cycle_trips
process_trips_los(cycle_trips, batch_size, save_path, "cycle_trips_los_gmap")

%%time
# RUN ONCE - Get LOS for walk_trips
process_trips_los(walk_trips, batch_size, save_path, "walk_trips_los_gmap")

### Join batch files together

train_trips_los = join_trips_los(len(train_trips), batch_size, save_path, "train_trips_los_gmap")
car_trips_los = join_trips_los(len(car_trips), batch_size, save_path, "car_trips_los_gmap")
walk_trips_los = join_trips_los(len(walk_trips), batch_size, save_path, "walk_trips_los_gmap")
cycle_trips_los = join_trips_los(len(cycle_trips), batch_size, save_path, "cycle_trips_los_gmap")
train_trips_los.shape, car_trips_los.shape, walk_trips_los.shape, cycle_trips_los.shape

full_trips_los = pd.concat([train_trips_los, car_trips_los, walk_trips_los, cycle_trips_los])
full_trips_los.shape

# Remove trips with TRANSIT_TIME == 0
full_trips_los = full_trips_los[(full_trips_los.TRANSIT_TIME != 0) & (full_trips_los.TRANSIT_LEN != 0)]
full_trips_los.drop(columns=["Unnamed: 0"], inplace=True)
full_trips_los.shape

full_trips_los.Mode.value_counts()

full_trips_los.head()

### Merge with trips_od to get full trip info

trips_od = pd.read_csv("../data/processed/trip_od_full.csv", low_memory=False)
cols = ['Mode', 'LINKMODE', 'TRIPPURP', 'DESTPURP1', 'TRIPTIME', 'TRAVTIME',
        'TRAVDOW', 'TRAVDATE', 'TRAVMONTH', 'TRAVYEAR', 'STARTHOUR', 'STARTIME', 
        'HomeSA1', 'CUMDIST','ORIGSA1', 'DESTSA1', 'ORIG_X', 'ORIG_Y', 'DEST_X', 'DEST_Y']
trips_od = trips_od.drop(columns=cols)

full_trips_los = full_trips_los.merge(trips_od, how='left', on='TRIPID')

# Save trips with LoS for further analysis
full_trips_los.to_csv("../data/processed/trip_gmap_los.csv", index=False)

full_trips_los.shape

full_trips_los.head().T

### Combine trip_gmap_los with trip_here_los

trip_gmap_los = pd.read_csv("../data/processed/trip_gmap_los.csv", low_memory=False)
trip_here_los = pd.read_csv("../data/processed/trip_here_los.csv", low_memory=False)
trip_gmap_los.shape, trip_here_los.shape

trip_los_full = trip_here_los.merge(trip_gmap_los[['TRIPID', 'TRANSIT_TIME', 'TRANSIT_LEN']], how='outer', on='TRIPID')
trip_los_full.dropna(inplace=True)
trip_los_full.shape

cols = ['TRIPID', 'Mode', 'LINKMODE', 'TRIPPURP', 'DESTPURP1', 'TRIPTIME', 'TRAVTIME', 'TRAVDOW', 'TRAVDATE', 'TRAVMONTH', 'TRAVYEAR',
 'STARTHOUR', 'STARTIME', 'HomeSA1', 'CUMDIST', 'ORIGSA1', 'DESTSA1', 'ORIG_X', 'ORIG_Y', 'DEST_X', 'DEST_Y', 'STARTMIN', 'START_DETAIL_TIME',
 'PERSID', 'TRIPNO', 'ARRHOUR', 'ARRTIME', 'WAITIME', 'ORIGSA3', 'DESTSA3', 'ORIG_CENTROID', 'DEST_CENTROID', 'HHID', 'PERSNO', 'WEEK',
 'MONTHOFBIRTH', 'YEAROFBIRTH', 'AGE', 'AGEGROUP', 'SEX', 'RELATIONSHIP', 'CARLICENCE', 'MBIKELICENCE', 'OTHERLICENCE', 'NOLICENCE',
 'FULLTIMEWORK', 'PARTTIMEWORK', 'CASUALWORK', 'ANYWORK', 'STUDYING', 'ACTIVITIES', 'MAINACT', 'WORKTYPE', 'EMPTYPE',
 'ANZSCO1', 'ANZSIC1', 'STARTPLACE', 'ANYSTOPS', 'NUMSTOPS', 'REASONCODE', 'MRTDOW', 'MRTINT', 'PERSINC', 'WHOFILLED', 'PROXY',
 'FILLDOW', 'FILLLAG', 'ADDITIONALTRAVEL', 'NEWSTOP', 'DISABLEDTRANS', 'TAXITRANS', 'WHEELCHAIRTRANS', 'ACCESSTRANS',
 'NONETRANS', 'NOTTOSAYTRANS', 'CYCLEDWORK', 'CYCLEDSHOPPING', 'CYCLEDEXERCISE', 'CYCLEDOTHER', 'NOCYCLED',
 'HOMESLA', 'HOMELGA', 'HomeSA2', 'HomeSA3', 'HomeSA4', 'HomeRegion_ASGS', 'HomeSubRegion_ASGS',
 'AgeGroup_RW', 'STOPS', 'ReportingPeriod', 'HHSIZE', 'VISITORS', 'DWELLTYPE', 'OWNDWELL', 'YEARSLIVED', 'MONTHSLIVED',
 'ADULTBIKES', 'KIDSBIKES', 'TOTALBIKES', 'PHONE', 'CARS', 'FOURWDS', 'UTES', 'VANS', 'TRUCKS', 'MBIKES', 'OTHERVEHS', 'TOTALVEHS',
 'HOMETAZ', 'YOUNGEST', 'AVEAGE', 'OLDEST', 'HHINC',
 'CAR_TIME', 'CAR_LEN', 'CYCLE_TIME', 'CYCLE_LEN', 'WALK_TIME', 'WALK_LEN', 'TRANSIT_TIME', 'TRANSIT_LEN']

trip_los_full = trip_los_full[cols]

trip_los_full.head()

trip_los_full.to_csv("../data/processed/trip_los_full.csv", index=False)
