==> notebooks/11_mode_choice_logit.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import biogeme.database as db
import biogeme.biogeme as bio
import biogeme.models as models
from biogeme.expressions import Beta

import config

### Get data

trips = pd.read_csv(config.TRIPS_1216_PRC, low_memory=False)
trips.shape

trips.head()

trips['Mode'].value_counts()

choice_mapping = {'Walk':1, 'Car':2, 'Train':3, 'Cycle':4}
trips['CHOICE'] = trips.Mode.map(choice_mapping)

# Only select these columns for MNL model
trips = trips[['TRIPTIME', 'TRAVTIME', 'WAITIME', 'CUMDIST','CHOICE']]

### Simple MNL model

database = db.Database('melbourne_mnl', trips)
globals().update(database.variables)

### Parameters to be estimatd:
ASC_WALK = Beta('ASC_WALK', 0, None, None, 0)
ASC_CAR = Beta('ASC_CAR', 0, None, None, 0)
ASC_TRAIN = Beta('ASC_TRAIN', 0, None, None, 0)
ASC_CYCLE = Beta('ASC_CYCLE', 0, None, None, 1)

B_TIME = Beta('B_TIME', 0, None, None, 0)

### Utility function
V_WALK = ASC_WALK + B_TIME * TRIPTIME
V_CAR = ASC_CAR + B_TIME * TRIPTIME
V_TRAIN = ASC_TRAIN + B_TIME * TRIPTIME
V_CYCLE = ASC_CYCLE + B_TIME * TRIPTIME

V = { 1: V_WALK, 2: V_CAR, 3: V_TRAIN, 4: V_CYCLE }

### Avaiability: Assume 1 for all alternatives
av = {1:1, 2:1, 3:1, 4:1}

### Create logit model
logprob = models.loglogit(V, av, CHOICE)
biogeme = bio.BIOGEME(database, logprob)

%%time
### Estimate paramaters
results = biogeme.estimate()

### Get results
results.getEstimatedParameters()

results.getGeneralStatistics()

results.getCorrelationResults()
